+++ tests/ArrayHelper/GroupTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Arrays\Tests\ArrayHelper;
 
-use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Arrays\ArrayHelper;
 
@@ -18,14 +17,14 @@ final class GroupTest extends TestCase
             ['id' => '345', 'data' => 'ghi'],
         ];
         $result = ArrayHelper::group($array, 'id');
-        $this->assertEquals(
+        self::assertEquals(
             [
                 '123' => [
-                  ['id' => '123', 'data' => 'abc']
+                    ['id' => '123', 'data' => 'abc']
                 ],
                 '345' => [
-                  ['id' => '345', 'data' => 'def'],
-                  ['id' => '345', 'data' => 'ghi'],
+                    ['id' => '345', 'data' => 'def'],
+                    ['id' => '345', 'data' => 'ghi'],
                 ],
             ],
             $result