+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.0.2 under development
 
 - New #91: Add method `ArrayHelper::group()` that groups the array according to a specified key (sagittaracc)
+- New #96: Add support iterable objects to `ArrayHelper::map()`, `ArrayHelper::index()`, `ArrayHelper::group()`,
+  `ArrayHelper::htmlEncode()` and `ArrayHelper::htmlDecode` (vjik)
 
 ## 1.0.1 February 10, 2021